Use Models' policies in JSONAPI routes.

+++ lib/JsonApi/Routes/CardsDelete.php
@@ -27,6 +27,10 @@ class CardsDelete extends JsonApiController
             throw new RecordNotFoundException();
         }
 
+        if ($this->cannot($request, 'delete', $resource)) {
+            throw new AuthorizationFailedException();
+        }
+
         $resource->delete();
 
         return $this->getCodeResponse(204);
